AiHelper/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
supabase
requests
pydub
